move error types and constants out to separate file
